tests and variable types

+++ moon-clock/_1_main.ino
@@ -1,4 +1,4 @@
-int mode;
+byte mode;
 
 void setup() {
