+++ moon-clock/_2_lights.ino
@@ -1,26 +1,25 @@
-
 void setupLight() {
   debugln("Setup light");
   
-  for(int i=0; i<DISPLAY_PRECISION; i++) {
+  for(byte i=0; i<DISPLAY_PRECISION; i++) {
     pinMode(PINS_START + i, OUTPUT);
   }
 
   lightAll(LOW);
 }
 
-void lightOne(int index, int state) {
+void lightOne(byte index, byte state) {
   digitalWrite(PINS_START + index, state);
 }
 
-void lightAll(int state) {
+void lightAll(byte state) {
   for(int i=0; i<DISPLAY_PRECISION; i++) {
     lightOne(i, state);
   }
 }
 
 void strobo() {
-  for(int i=0; i<10; i++) {
+  for(byte i=0; i<10; i++) {
     lightAll(HIGH);
     delay(50);
     lightAll(LOW);
